requirement.txt
python-dotenv==0.21.0
pymongo==4.2.0
SQLAlchemy==1.4.41
pandas==1.4.4
requests==2.28.1